requirements.txt
streamlit
importlib-metadata
pdfminer.six
PyPDF2
openai
